pyproject.toml
[project]
name = "fastapi-eccomerce"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic==1.15.1",
    "annotated-types==0.7.0",
    "anyio==4.8.0",
    "asyncpg==0.30.0",
    "certifi==2025.1.31",
    "charset-normalizer==3.4.1",
    "click==8.1.8",
    "fastapi==0.115.11",
    "greenlet>=3.1.1",
    "h11==0.14.0",
    "idna==3.10",
    "mako==1.3.9",
    "markupsafe==3.0.2",
    "naked==0.1.32",
    "pydantic==2.10.6",
    "pydantic-core==2.27.2",
    "python-slugify==8.0.4",
    "pyyaml==6.0.2",
    "requests==2.32.3",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.39",
    "starlette==0.46.1",
    "status==0.2.5",
    "text-unidecode==1.3",
    "typing-extensions==4.12.2",
    "urllib3==2.3.0",
    "uv==0.6.6",
    "uvicorn==0.34.0",
]
